+++ ibis/backends/base_sql/__init__.py
@@ -9,11 +9,11 @@ import itertools
 import math
 
 import ibis
+import ibis.backends.base_sqlalchemy.compiler as comp
 import ibis.common.exceptions as com
 import ibis.expr.datatypes as dt
 import ibis.expr.operations as ops
 import ibis.expr.types as ir
-import ibis.sql.compiler as comp
 from ibis.impala import identifiers
