+++ ibis/backends/base_sqlalchemy/alchemy.py
@@ -13,6 +13,7 @@ from sqlalchemy.dialects.sqlite.base import SQLiteDialect
 from sqlalchemy.engine.interfaces import Dialect as SQLAlchemyDialect
 
 import ibis
+import ibis.backends.base_sqlalchemy.compiler as comp
 import ibis.common.exceptions as com
 import ibis.expr.analysis as L
 import ibis.expr.datatypes as dt
@@ -20,11 +21,15 @@ import ibis.expr.operations as ops
 import ibis.expr.schema as sch
 import ibis.expr.types as ir
 import ibis.expr.window as W
-import ibis.sql.compiler as comp
 import ibis.sql.transforms as transforms
 import ibis.util as util
+from ibis.backends.base_sqlalchemy.compiler import (
+    Dialect,
+    Select,
+    TableSetFormatter,
+    Union,
+)
 from ibis.client import Database, Query, SQLClient
-from ibis.sql.compiler import Dialect, Select, TableSetFormatter, Union
 
 geospatial_supported = False
 try:
