+++ ibis/backends/spark/compiler.py
@@ -24,12 +24,12 @@ import itertools
 import math
 
 import ibis
+import ibis.backends.base_sqlalchemy.compiler as comp
 import ibis.common.exceptions as com
 import ibis.expr.datatypes as dt
 import ibis.expr.operations as ops
 import ibis.expr.rules as rlz
 import ibis.expr.types as ir
-import ibis.sql.compiler as comp
 import ibis.util as util
 from ibis.backends import base_sql
 from ibis.backends.base_sql import (
